Separated controllers,models,views
