+++ fbadmin/inc/config.php
@@ -24,5 +24,13 @@ require_once 'view/View.php';
 require_once 'modules/articles/Article.php';
 require_once 'modules/User.php';
 
+require_once 'controller/ArticleController.php';
+require_once 'model/ArticleModel.php';
+require_once 'view/ArticleView.php';
+
+require_once 'controller/UserController.php';
+require_once 'model/UserModel.php';
+require_once 'view/UserView.php';
+
 require_once 'modules/fbvars.php';
 require_once 'modules/auth.php';
