+++ fbadmin/inc/model/UserModel.php
@@ -1,0 +1,14 @@
+<?php
+namespace FreeBlog\Admin\Model;
+
+/**
+ * Description of UserModel
+ *
+ * @author myrmidex
+ */
+class UserModel extends Model 
+{
+
+    
+    
+}
